+++ Frontend/src/lib/components/index.ts
@@ -1,9 +1,11 @@
 export {default as Header} from './Header.svelte';
 export {default as Footer} from './Footer.svelte';
-export {default as Cheveron} from './Cheveron.svelte';
 export {default as Notification} from './Notification.svelte';
 
+export {default as CheveronImg} from './icons/Cheveron.svelte';
+export {default as QuestionImg} from './icons/Question.svelte';
+
 export {default as BackgroundBeams} from './aceternity/BackgroundBeams.svelte';
+export {default as CardContainer} from './aceternity/CardContainer.svelte';
 export {default as CardBody} from './aceternity/CardBody.svelte';
 export {default as CardItem} from './aceternity/CardItem.svelte';
-export {default as CardContainer} from './aceternity/CardContainer.svelte';