+++ Frontend/src/routes/+error.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import {page} from '$app/stores';
+  import {CheveronImg} from '$components';
   import {circuitPattern, itCrowdGif} from '$images';
-  import {Cheveron} from '$components';
 </script>
 
 <svelte:head>
@@ -14,7 +14,7 @@
   <img alt={$page.error?.message} src={itCrowdGif} />
   <h1 class="text-7xl">Page Not Found.</h1>
   <p>Oops, looks like you hit a dead end! That's not a big deal, Let's get you back on track.</p>
-  <a href="/"><button>Go Home<Cheveron /></button></a>
+  <a href="/"><button>Go Home<CheveronImg /></button></a>
 </section>
 
 <style lang="postcss">
@@ -23,7 +23,7 @@
   }
 
   img {
-    @apply w-auto rounded-sm border border-neutral-300 shadow-2xl;
+    @apply w-auto rounded-sm border border-neutral-400 shadow-2xl;
   }
 
   p {
@@ -31,6 +31,6 @@
   }
 
   button {
-    @apply flex items-center gap-2 px-6 py-5 text-2xl;
+    @apply flex items-center gap-3 px-6 py-5 text-2xl;
   }
 </style>