+++ CHANGELOG.md
@@ -5,13 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.10] - 2024-10-11
+
+### Added
+
+- Add pricing section on homepage
+- Add products section on homepage
+
+### Changed
+
+- Change home section text
+- Change icon images to components
+
 ## [0.0.9] - 2024-10-10
 
 ### Added
 
-- Add reflection section on homepage
-- Add pricing section on homepage
 - Add no script client support
+- Add reflection section on homepage
 - Create highlight and underline effect on homepage
 
 ### Changed